Revise📖 & Features🌍 & Fix issue🔧 1. 📖Added `system_tray` `window_manager` package, updated some packages 2. 🌍Enabled Windows version of Mercurius 3. 📖Added more try-catch to avoid crash 4. 🔧Fixed issue that occurs at deleting last of diary
Former-commit-id: 44d51881b6cf4cd7723608cba1c2e4fec859af54

+++ lib/widgets/diary/editor/diary_editor_toolbar_image_button_widget.dart
@@ -53,7 +53,9 @@ class DiaryEditorToolbarImageButtonWidget extends QuillIconButton {
           String targetFilePath = '$path/image/${pickedFile.name}';
 
           await XFile(sourceFilePath).saveTo(targetFilePath);
-          await File(sourceFilePath).delete();
+          if (Platform.isAndroid) {
+            await File(sourceFilePath).delete();
+          }
           _insertImage(controller, targetFilePath);
         }
         break;
